chore: add new PNG files to data/png_txt/figs_0 and figs_1 directories
This commit introduces a large number of new PNG files to the data/png_txt/figs_0 and figs_1 directories. These files are part of a dataset update and will be used for further processing and analysis. No functional changes are made in this commit; it is purely an addition of new resources.